feat: changed docker settings, more tests
