+++ docker-compose.yml
@@ -2,7 +2,7 @@ version: '3.8'
 
 services:
   app:
-    image: 'docker-spring-boot-postgres:latest'
+    image: 'aptmgr.0.1'
     build:
       context: .
     container_name: aptmgr
