+++ src/main/java/org/example/ApplicationEntry.java
@@ -1,0 +1,14 @@
+package org.example;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
+import org.springframework.boot.actuate.autoconfigure.security.servlet.ManagementWebSecurityAutoConfiguration;
+
+@SpringBootApplication(exclude = { SecurityAutoConfiguration.class, ManagementWebSecurityAutoConfiguration.class })
+public class ApplicationEntry {
+    public static void main(String[] args) {
+        SpringApplication.run(ApplicationEntry.class, args);
+        
+    }
+}
