--- src/main/java/org/example/SecurityConfig/ApplicationConfiguration.java
@@ -1,45 +1,0 @@
-package org.example.SecurityConfig;
-
-import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
-import org.example.user.UserRepository;
-import org.example.util.error.EntityNotFoundException;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-
-@Configuration
-@AllArgsConstructor
-public class ApplicationConfiguration {
-    private final UserRepository userRepository;
-
-    @Bean
-    UserDetailsService userDetailsService() {
-        return username -> userRepository.findByName(username)
-                .orElseThrow(() -> new EntityNotFoundException(String.format("User with name=%s was not found", username),
-                        "The required object was not found."));
-    }
-
-    @Bean
-    BCryptPasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-    @Bean
-    public AuthenticationManager authenticationManager(AuthenticationConfiguration config) throws Exception {
-        return config.getAuthenticationManager();
-    }
-
-    @Bean
-    AuthenticationProvider authenticationProvider() {
-        DaoAuthenticationProvider authProvider = new DaoAuthenticationProvider();
-        authProvider.setUserDetailsService(userDetailsService());
-        authProvider.setPasswordEncoder(passwordEncoder());
-        return authProvider;
-    }
-}
