--- src/main/java/org/example/SecurityConfig/AuthenticationService.java
@@ -1,10 +1,0 @@
-package org.example.SecurityConfig;
-
-import org.example.user.dto.UserAuthDto;
-import org.example.user.dto.UserDto;
-import org.example.util.auth.LoginResponse;
-
-public interface AuthenticationService {
-    UserDto signup(UserAuthDto userAuthDto);
-    LoginResponse authenticate(UserAuthDto userAuthDto);
-}
